+++ test/cookbooks/loggly-rsyslog_test/metadata.rb
@@ -1,8 +1,0 @@
-name             "loggly_rsyslog_ng_test"
-maintainer       'Kostiantyn Lysenko'
-maintainer_email 'gshaud@gmail.com'
-license          'Apache 2.0'
-description      "Used to test the Loggly rsyslog cookbook"
-version          "1.0.0"
-
-depends "loggly_rsyslog_ng"